+++ week4-task/src/types/search.types.ts
@@ -1,0 +1,8 @@
+export interface SearchNicknameResponse {
+  success: boolean;
+  code: string;
+  message: string;
+  data: {
+    nicknameList: string[];
+  } | null;
+}
