+++ week4-task/src/types/user.types.ts
@@ -1,0 +1,10 @@
+export interface NicknameData {
+  nickname: string;
+}
+
+export interface ApiResponse {
+  success: boolean;
+  code: string;
+  message: string;
+  data: NicknameData | null;
+}
